fix(counter): use u64 saturating arithmetic and derive Copy for Count (#21)
intent(counter): close two related robustness/perf findings in one PR
since both touch the same Count struct definition (#14, #8).
decision(counter): widen Count fields to u64 with saturating_add rather
than checked/overflowing variants — a saturated total is visibly
"pegged" instead of silently wrapping or panicking, and u64 makes the
limit uniform across 32-bit and 64-bit targets.
decision(counter): derive Copy on Count (32 bytes, four integers) and
drop the now-redundant .clone() calls at aggregation call sites instead
of adding a Sum<&Count> impl, since Copy also simplifies future callers.

+++ src/output.rs
@@ -6,7 +6,7 @@ pub enum OutputKind {
     Directory(usize),
 }
 
-pub fn format_number(n: usize) -> String {
+pub fn format_number(n: u64) -> String {
     n.to_string()
         .as_bytes()
         .rchunks(3)
